+++ handlers/nonce/nonce.go
@@ -1,91 +1,114 @@
 package nonce
 
 import (
-	"context"
-	"math/big"
+	"strconv"
 
-	"github.com/ethereum/go-ethereum/common"
 	log "github.com/sirupsen/logrus"
 	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/engine"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/errors"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/types/envelope"
 	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/types/ethereum"
+	"gitlab.com/ConsenSys/client/fr/core-stack/service/ethereum.git/ethclient"
 	"gitlab.com/ConsenSys/client/fr/core-stack/service/nonce.git/nonce"
 )
 
-// GetNonceFunc should return an effective nonce for calibration (usually retrieved from an EThereum)
-type GetNonceFunc func(ctx context.Context, chainID *big.Int, a common.Address) (uint64, error)
-
 // Handler creates and return an handler for nonce
-func Handler(nc nonce.Nonce, getChainNonce GetNonceFunc) engine.HandlerFunc {
+func Nonce(nm nonce.Attributor, ec ethclient.ChainStateReader) engine.HandlerFunc {
 	return func(txctx *engine.TxContext) {
 		// Retrieve chainID and sender address
-		chainID := txctx.Envelope.GetChain().ID()
-
-		a := txctx.Envelope.GetFrom().Address()
-
+		chainID, sender := txctx.Envelope.GetChain().ID(), txctx.Envelope.GetFrom().Address()
 		txctx.Logger = txctx.Logger.WithFields(log.Fields{
-			"tx.sender":   a.Hex(),
+			"tx.sender":   sender.Hex(),
 			"chain.id":    chainID.String(),
 			"metadata.id": txctx.Envelope.GetMetadata().GetId(),
 		})
 
-		// Get the lock for chainID and sender address
-		lockSig, err := nc.Lock(chainID, &a)
-		if err != nil {
-			e := txctx.AbortWithError(err).ExtendComponent(component)
-			txctx.Logger.WithError(e).Errorf("nonce: could not acquire nonce lock")
-			return
-		}
-		defer func() {
-			err = nc.Unlock(chainID, &a, lockSig)
+		// Nonce to attribute to tx
+		var n uint64
+		// Compute nonce key for nonce manager procesing
+		nonceKey := string(txctx.In.Key())
+
+		// First check if signal for recovering nonce
+		if v, ok := txctx.Envelope.GetMetadataValue("nonce.recovering.expected"); ok {
+			expectedNonce, err := strconv.ParseUint(v, 10, 64)
 			if err != nil {
 				e := txctx.AbortWithError(err).ExtendComponent(component)
-				txctx.Logger.WithError(e).Warnf("nonce: could not release nonce lock")
+				txctx.Logger.WithError(e).Errorf("nonce: could not extract metadata")
 				return
 			}
-		}()
 
-		// Retrieve nonce
-		n, inCache, err := nc.Get(chainID, &a)
-		if err != nil {
-			e := txctx.AbortWithError(err).ExtendComponent(component)
-			txctx.Logger.WithError(e).Errorf("nonce: could not get nonce from cache")
-			return
-		}
+			txctx.Logger.WithFields(log.Fields{
+				"nonce.expected": expectedNonce,
+			}).Warnf("nonce: recalibrate nonce following recovery signal")
 
-		// If nonce was not in cache, we calibrate it by reading nonce from chain
-		if !inCache {
-			txctx.Logger.Debugf("nonce: not in cache, get from chain")
-			n, err = getChainNonce(txctx.Context(), chainID, a)
+			n = expectedNonce
+
+			// Reset signal field
+			delete(txctx.Envelope.GetMetadata().GetExtra(), "nonce.recovering.expected")
+		} else {
+			// No signal for nonce recovery
+			// Retrieve last attributed nonce from nonce manager
+			lastAttributed, ok, err := nm.GetLastAttributed(nonceKey)
 			if err != nil {
 				e := txctx.AbortWithError(err).ExtendComponent(component)
-				txctx.Logger.WithError(e).Errorf("nonce: could not get nonce from chain")
+				txctx.Logger.WithError(e).Errorf("nonce: could not retrieve last attributed nonce")
 				return
+			}
+
+			// If no nonce is available in nonce manager
+			// we calibrate by querying chain
+			if !ok {
+				txctx.Logger.Debugf("nonce: not in cache, get from chain")
+
+				// Retrieve nonce from chain
+				pendingNonce, err := ec.PendingNonceAt(txctx.Context(), chainID, sender)
+				if err != nil {
+					e := txctx.AbortWithError(err).ExtendComponent(component)
+					txctx.Logger.WithError(e).Errorf("nonce: could not read nonce from chain")
+					return
+				}
+
+				n = pendingNonce
+			} else {
+				n = lastAttributed + 1
 			}
 		}
 
-		// Set Nonce value on Envelope
-		if txctx.Envelope.GetTx() == nil {
-			txctx.Envelope.Tx = &ethereum.Transaction{TxData: &ethereum.TxData{}}
-		} else if txctx.Envelope.GetTx().GetTxData() == nil {
-			txctx.Envelope.Tx.TxData = &ethereum.TxData{}
-		}
-		txctx.Envelope.GetTx().GetTxData().SetNonce(n)
+		// Set nonce
+		setNonce(txctx.Envelope, n)
 		txctx.Logger = txctx.Logger.WithFields(log.Fields{
 			"tx.nonce": n,
 		})
-		txctx.Logger.Debugf("nonce: nonce set")
 
-		// Execute pending handlers (note that we do not release lock while executing pending handlers)
+		// Execute pending handlers
 		txctx.Next()
 
-		// Increment nonce in Manager
-		// TODO: we should ensure pending handlers have correctly executed before incrementing
-		err = nc.Set(chainID, &a, n+1)
-		if err != nil {
-			// TODO: handle error
-			e := txctx.AbortWithError(err).ExtendComponent(component)
-			txctx.Logger.WithError(e).Errorf("nonce: could not set nonce on cache")
-			return
+		// If pending handlers executed correctly we increment nonce
+		if len(txctx.Envelope.GetErrors()) == 0 {
+			// Increment nonce
+			err := nm.SetLastAttributed(nonceKey, n)
+			if err != nil {
+				// An error here means that we probably lost connection with NonceManager underlying cache.
+				// TODO: A retry strategy should be implemented on the nonce manager to make this scenario rare
+				//
+				// At this point pending handlers have executed correctly
+				// meaning that envelope has already been successfully produced to the next topic
+				//
+				// So we log the error and process next envelope
+				e := errors.FromError(err).ExtendComponent(component)
+				txctx.Logger.WithError(e).Errorf("nonce: could not set nonce on cache")
+			}
 		}
 	}
+}
+
+func setNonce(e *envelope.Envelope, n uint64) {
+	// Initialiaze Transaction on envelope if needed
+	if e.GetTx() == nil {
+		e.Tx = &ethereum.Transaction{TxData: &ethereum.TxData{}}
+	} else if e.GetTx().GetTxData() == nil {
+		e.Tx.TxData = &ethereum.TxData{}
+	}
+	// Set transaction nonce
+	e.GetTx().GetTxData().SetNonce(n)
 }